Pull out database driver into one place
