+++ src/test/scala/ChatTests.scala
@@ -3,7 +3,7 @@ import org.scalatra.test.scalatest._
 import org.scalatest.FunSuiteLike
 import akka.actor.{ActorSystem}
 import akka.testkit.{TestKit, ImplicitSender}
-import slick.driver.H2Driver.api._
+import org.playarimaa.server.DatabaseConfig.driver.api._
 import scala.util.{Try, Success, Failure}
 import scala.concurrent.ExecutionContext
 import com.typesafe.config.ConfigFactory
@@ -18,7 +18,7 @@ class ChatServletTests(_system: ActorSystem) extends TestKit(_system) with Scala
 
   def this() = this(ActorSystem("ChatTests"))
 
-  override def afterAll {
+  override def afterAll : Unit = {
     TestKit.shutdownActorSystem(system)
   }
 
@@ -42,7 +42,7 @@ class ChatServletTests(_system: ActorSystem) extends TestKit(_system) with Scala
   val cryptEC: ExecutionContext = mainEC
   val actorEC: ExecutionContext = actorSystem.dispatcher
   val serverInstanceID: Long = System.currentTimeMillis
-  val db = ArimaaServerInit.createDB("h2memchat")
+  val db = DatabaseConfig.createDB("h2memchat")
   val scheduler = actorSystem.scheduler
   val emailer = new Emailer(siteName,siteAddress,smtpHost,smtpPort,smtpAuth,noReplyAddress,helpAddress)(mainEC)
   val accounts = new Accounts(db,scheduler)(mainEC)
